# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kisara API",
    "description": "API documentation untuk layanan kisara.my.id (generated otomatis dari route schemas).",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/message/{link_id}": {
      "post": {
        "summary": "Create message",
        "tags": [
          "Message"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message_content": {
                    "minLength": 1,
                    "maxLength": 500,
                    "type": "string"
                  }
                },
                "required": [
                  "message_content"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "link_id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number",
                      "enum": [
                        201
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "message_content": {
                          "type": "string"
                        },
                        "created_at": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "message_content",
                        "created_at"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "data",
                    "metadata"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {}
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "errors",
                    "metadata"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get messages by link",
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "schema": {
              "default": "desc",
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "asc"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "desc"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sortBy",
            "required": false
          },
          {
            "schema": {
              "default": "desc",
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "asc"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "desc"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort_by",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "link_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number",
                      "enum": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total_records": {
                          "type": "integer"
                        },
                        "total_pages": {
                          "type": "integer"
                        },
                        "author": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            },
                            "profile_url": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "name",
                            "role",
                            "profile_url"
                          ]
                        },
                        "comments": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "message_content": {
                                "type": "string"
                              },
                              "like_by_creator": {
                                "type": "boolean"
                              },
                              "created_at": {
                                "format": "date-time",
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "message_content",
                              "like_by_creator",
                              "created_at"
                            ]
                          }
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total_records",
                        "total_pages",
                        "author",
                        "comments"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "data",
                    "metadata"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {}
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "errors",
                    "metadata"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/message/{link_id}/{message_id}/{reply_id}": {
      "delete": {
        "summary": "Delete reply",
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "link_id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "message_id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "reply_id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number",
                      "enum": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "data",
                    "metadata"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.kisara.my.id",
      "description": "Production"
    }
  ],
...
}